Use macros to generate object hierarchy boilerplate.
Also, extract method signature computation into the methods module.
